requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
pillow==10.1.0
yt-dlp==2023.12.30
wavelink==1.3.5
PyNaCl==1.5.0
pytz==2023.3.post1
psutil==5.9.7
flask==3.0.0
